Overlay scene lights on all lights
This patch creates a SceneLight object as a child of Light so that such
objects can be aware of whether they're actually part of a scene or
merely for display.

Scenes will only store the lights they control; all other lights will be
left alone when setting (which was the prior behavior).